moving byteutils and notify system to their own files
