+++ Assets/Mirage/Runtime/Transport/SocketLayer/AckSystem.cs
@@ -1,72 +1,8 @@
 using System;
-using System.Collections.Generic;
 using UnityEngine.Assertions;
 
 namespace Mirage.SocketLayer
 {
-    internal static class ByteUtils
-    {
-        public static void WriteByte(byte[] buffer, ref int offset, byte value)
-        {
-            buffer[offset] = value;
-            offset++;
-        }
-
-        public static byte ReadByte(byte[] buffer, ref int offset)
-        {
-            byte a = buffer[offset];
-            offset++;
-
-            return a;
-        }
-
-
-        public static void WriteUShort(byte[] buffer, ref int offset, ushort value)
-        {
-            buffer[offset] = (byte)value;
-            offset++;
-            buffer[offset] = (byte)(value >> 8);
-            offset++;
-        }
-
-        public static ushort ReadUShort(byte[] buffer, ref int offset)
-        {
-            ushort a = buffer[offset];
-            offset++;
-            ushort b = buffer[offset];
-            offset++;
-
-            return (ushort)(a | (b << 8));
-        }
-
-
-        public static void WriteUInt(byte[] buffer, ref int offset, uint value)
-        {
-            buffer[offset] = (byte)value;
-            offset++;
-            buffer[offset] = (byte)(value >> 8);
-            offset++;
-            buffer[offset] = (byte)(value >> 16);
-            offset++;
-            buffer[offset] = (byte)(value >> 24);
-            offset++;
-        }
-
-        public static uint ReadUInt(byte[] buffer, ref int offset)
-        {
-            uint a = buffer[offset];
-            offset++;
-            uint b = buffer[offset];
-            offset++;
-            uint c = buffer[offset];
-            offset++;
-            uint d = buffer[offset];
-            offset++;
-
-            return a | (b << 8) | (c << 16) | (d << 24);
-        }
-    }
-
     struct ReceivedPacket
     {
         public ushort sequence;
@@ -224,89 +160,4 @@ namespace Mirage.SocketLayer
         }
     }
 
-    internal class NotifySystem
-    {
-        readonly AckSystem ackSystem;
-        readonly float timeout;
-
-
-        const int MaxSentQueue = 512;
-        Queue<NotifyToken> sent = new Queue<NotifyToken>(MaxSentQueue);
-
-        public NotifySystem(IRawConnection connection, float timeout, Time time)
-        {
-            ackSystem = new AckSystem(connection, time);
-            this.timeout = timeout;
-        }
-
-        public NotifyToken Send(byte[] packet)
-        {
-            if (sent.Count >= MaxSentQueue)
-            {
-                throw new InvalidOperationException("Sent queue is full");
-            }
-
-            ackSystem.Send(packet);
-
-            // todo use pool to stop allocations
-            return new NotifyToken();
-        }
-        public void Receive(byte[] packet)
-        {
-            //todo get messages from pecket and invoke
-            ReceivedPacket received = ackSystem.Receive(packet);
-            while (sent.Count > 0)
-            {
-                NotifyToken next = sent.Peek();
-
-                int distance = (int)ackSystem.sequencer.Distance(next.Sequence, received.receivedSequence);
-
-                // posititve distance means next is sent after last ack, so nothing to ack yet
-                if (distance > 0)
-                    return;
-
-                // negative distance means it should have been acked, or mark it as lost
-                sent.Dequeue();
-
-                const int maskSize = sizeof(uint) * 8;
-                int posDistance = -distance;
-                // if distance above size then it is outside of mask, so set as lost
-
-                bool outsideOfMask = posDistance > maskSize;
-
-                uint ackBit = 1u << posDistance;
-                bool notInMask = (received.receivedMask & ackBit) == 0u;
-
-                // todo clean this code up with multiple methods
-                bool lost = outsideOfMask || notInMask;
-
-                next.Notify(!lost);
-            }
-        }
-
-        public void Update()
-        {
-            ackSystem.Update();
-        }
-    }
-
-    public class NotifyToken
-    {
-        public event Action Delivered;
-        public event Action Lost;
-
-        bool notified;
-        internal ushort Sequence;
-
-        internal void Notify(bool delivered)
-        {
-            if (notified) throw new InvalidOperationException("this token as already been notified");
-            notified = true;
-
-            if (delivered)
-                Delivered.Invoke();
-            else
-                Lost.Invoke();
-        }
-    }
 }
