+++ Assets/Mirage/Runtime/Transport/SocketLayer/ByteUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bacac50e5440be241992174727872eb2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
