+++ Assets/Mirage/Runtime/Transport/SocketLayer/NotifySystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd550cfa14fbd67458170175625448bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
